Add cleanConsumerOffsetList Test
