--- common/src/main/java/org/apache/rocketmq/common/protocol/body/OffsetWrapper.java
@@ -1,44 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.rocketmq.common.protocol.body;
-
-import org.apache.rocketmq.common.DataVersion;
-import org.apache.rocketmq.remoting.protocol.RemotingSerializable;
-
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-
-public class OffsetWrapper extends RemotingSerializable {
-
-    private ConcurrentMap<String/* topic@group */, ConcurrentMap<Integer, Long>> offsetTable =
-            new ConcurrentHashMap<String, ConcurrentMap<Integer, Long>>(512);
-
-    private DataVersion dataVersion = new DataVersion();
-
-    public  ConcurrentMap<String/* topic@group */, ConcurrentMap<Integer, Long>> getOffsetTable() {
-        return offsetTable;
-    }
-
-    public void setOffsetTable(ConcurrentMap<String/*topi@group */,ConcurrentMap<Integer,Long>> offsetTable) {
-        this.offsetTable = offsetTable;
-    }
-
-    public void setDataVersion(DataVersion dataVersion) {
-        this.dataVersion = dataVersion;
-    }
-}
